refactor(shared): extract shared blob-download helper (DRY)
